requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
Django==5.0.1
django-redis==5.4.0
djangorestframework==3.14.0
future==0.18.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mysqlclient==2.2.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2024.1
redis==5.0.1
serializers==0.2.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
zope.interface==6.1
